Remove dead code and improve codebase hygiene
- Delete unused forms/ directory (7 files)
- Delete unused lib components (FormHook, ContactForm, FormWrapper, etc.)
- Delete unused page components (Test.tsx, GetInvolved.tsx, TemplateFormPage.tsx)
- Rename test.css to pages.css for clarity
- Rename getInvolvedRedux.tsx to GetInvolved.tsx
- Clean up unused imports across components
- Merge nonHoverFooter.css into footer.css
- Add CLAUDE.md for AI-assisted maintenance
