+++ kxcj-lp-front-end/src/SubmitPSA.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useRef } from "react";
 import { useLocation } from "react-router-dom";
 import emailjs from "@emailjs/browser";
-import "./styles/test.css";
+import "./styles/pages.css";
 import Scroll from "./assets/Scrolls/Scroll_p.png";
 import SideButtons from "./lib/SideButtons";
