--- kxcj-lp-front-end/src/forms/PSAFormFinal.css
@@ -1,65 +1,0 @@
-.formContent-psa {
-    display: grid;
-    gap: 1rem .5rem;
-    justify-content: flex-start;
-    grid-template-columns: auto minmax(auto, 400px);
-}
-
-.form-psa form {
-    display: flex;
-    flex-direction: column;
-    align-items: left;
-    gap: 20px;
-}
-
-form-label {
-    display: block;
-    margin-bottom: 10px;
-    font-weight: bold;
-  }
-  
-  .form-input,
-  .form-textarea {
-    width: 100%;
-    padding: 10px;
-    font-size: 16px;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-    transition: border-color 0.3s ease;
-  }
-  
-  .form-input:focus,
-  .form-textarea:focus {
-    outline: none;
-    border-color: #4c8bf5;
-  }
-  
-  .form-textarea {
-    width: 100%;
-    padding: 10px;
-    font-size: 16px;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-    resize: vertical;
-    transition: border-color 0.3s ease;
-  }
-  
-  /* Adjust the height of the textarea to make it smaller */
-  .form-textarea {
-    height: 120px;
-  }
-  
-  .form-submit {
-    background-color: #4c8bf5;
-    color: #fff;
-    border: none;
-    padding: 12px 20px;
-    font-size: 16px;
-    border-radius: 4px;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-  }
-  
-  .form-submit:hover {
-    background-color: #3469b7;
-  }
