--- kxcj-lp-front-end/src/lib/FormWrapper.tsx
@@ -1,15 +1,0 @@
-import { ReactNode } from "react";
-import "../styles/form.css";
-type FormWrapperProps = {
-  title: string;
-  children: ReactNode;
-};
-
-export function FormWrapper({ title, children }: FormWrapperProps) {
-  return (
-    <div className="form">
-      <h2 className="formHeading">{title}</h2>
-      <div className="formContent">{children}</div>
-    </div>
-  );
-}
