--- kxcj-lp-front-end/src/styles/formPages.css
@@ -1,44 +1,0 @@
-.top-level-container {
-    height: 90%;
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: center;
-    padding: 60px;
-    overflow: hidden;
-    
-}
-
-.glass-container {
-    max-width: 800px;
-    text-align: center;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.glass-container h1{
-font-size: 3rem;    
-}
-
-.glass-container {
-    height: 1000px;
-}
-
-.glass-container img{
-    width: 900px;   
-    margin-left: 170px;
-    
-    }
-
-    .sideBut {
-        margin-top: 300px;
-        padding-left: 100px;
-    }
-
-    .form-container {
-        
-    }
-    .scroll img {
-        width: 800px;
-    }
